+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/directconnectivity/rntbd/RntbdServiceEndpoint.java
@@ -91,6 +91,7 @@ public final class RntbdServiceEndpoint implements RntbdEndpoint {
 
     private final RntbdConnectionStateListener connectionStateListener;
     private final URI serviceEndpoint;
+    private final RntbdDurableEndpointMetrics durableMetrics;
     private String lastFaultInjectionRuleId;
     private Instant lastFaultInjectionTimestamp;
 
@@ -106,8 +107,10 @@ public final class RntbdServiceEndpoint implements RntbdEndpoint {
         final URI physicalAddress,
         final ClientTelemetry clientTelemetry,
         final RntbdServerErrorInjector faultInjectionInterceptors,
-        final URI serviceEndpoint) {
+        final URI serviceEndpoint,
+        final RntbdDurableEndpointMetrics durableMetrics) {
 
+        this.durableMetrics = durableMetrics;
         this.serverKey = RntbdUtils.getServerKey(physicalAddress);
         this.serviceEndpoint = serviceEndpoint;
 
@@ -145,7 +148,8 @@ public final class RntbdServiceEndpoint implements RntbdEndpoint {
                 config,
                 clientTelemetry,
                 this.connectionStateListener,
-                faultInjectionInterceptors);
+                faultInjectionInterceptors,
+                durableMetrics);
 
         if (clientTelemetry != null &&
             clientTelemetry.isClientMetricsEnabled() &&
@@ -216,20 +220,9 @@ public final class RntbdServiceEndpoint implements RntbdEndpoint {
         return this.channelPool.channelsAcquiredMetrics();
     }
 
-    /**
-     * @return approximate number of acquired channels.
-     */
     @Override
-    public int totalChannelsAcquiredMetric() {
-        return this.channelPool.totalChannelsAcquiredMetrics();
-    }
-
-    /**
-     * @return approximate number of closed channels.
-     */
-    @Override
-    public int totalChannelsClosedMetric() {
-        return this.channelPool.totalChannelsClosedMetrics();
+    public RntbdDurableEndpointMetrics durableEndpointMetrics() {
+        return this.durableMetrics;
     }
 
     /**
@@ -344,6 +337,7 @@ public final class RntbdServiceEndpoint implements RntbdEndpoint {
     public void close() {
         if (this.closed.compareAndSet(false, true)) {
             this.provider.evict(this);
+            this.durableMetrics.clearEndpoint(this);
             this.channelPool.close();
         }
     }
@@ -622,6 +616,7 @@ public final class RntbdServiceEndpoint implements RntbdEndpoint {
         private final AtomicBoolean closed;
         private final Config config;
         private final ConcurrentHashMap<String, RntbdEndpoint> endpoints;
+        private final ConcurrentHashMap<String, RntbdDurableEndpointMetrics> durableMetrics;
         private final EventLoopGroup eventLoopGroup;
         private final AtomicInteger evictions;
         private final RntbdEndpointMonitoringProvider monitoring;
@@ -661,6 +656,7 @@ public final class RntbdServiceEndpoint implements RntbdEndpoint {
 
             this.eventLoopGroup = this.getEventLoopGroup(options);
             this.endpoints = new ConcurrentHashMap<>();
+            this.durableMetrics = new ConcurrentHashMap<>();
             this.evictions = new AtomicInteger();
             this.closed = new AtomicBoolean();
             this.clientTelemetry = clientTelemetry;
@@ -728,15 +724,30 @@ public final class RntbdServiceEndpoint implements RntbdEndpoint {
 
         @Override
         public RntbdEndpoint createIfAbsent(final URI serviceEndpoint, final URI physicalAddress) {
-            return endpoints.computeIfAbsent(physicalAddress.getAuthority(), authority -> new RntbdServiceEndpoint(
-                this,
-                this.config,
-                this.eventLoopGroup,
-                this.requestTimer,
-                physicalAddress,
-                this.clientTelemetry,
-                this.serverErrorInjector,
-                serviceEndpoint));
+            return endpoints.computeIfAbsent(
+                physicalAddress.getAuthority(),
+                authority -> {
+
+                    RntbdDurableEndpointMetrics durableEndpointMetrics = durableMetrics.computeIfAbsent(
+                        authority,
+                        ignoreMe -> new RntbdDurableEndpointMetrics()
+                    );
+
+                    RntbdServiceEndpoint endpoint = new RntbdServiceEndpoint(
+                        this,
+                        this.config,
+                        this.eventLoopGroup,
+                        this.requestTimer,
+                        physicalAddress,
+                        this.clientTelemetry,
+                        this.serverErrorInjector,
+                        serviceEndpoint,
+                        durableEndpointMetrics);
+
+                    durableEndpointMetrics.setEndpoint(endpoint);
+
+                    return endpoint;
+                });
         }
 
         @Override