+++ sdk/hybridcontainerservice/azure-resourcemanager-hybridcontainerservice/src/main/java/com/azure/resourcemanager/hybridcontainerservice/models/HybridIdentityMetadatas.java
@@ -16,7 +16,7 @@ public interface HybridIdentityMetadatas {
      * <p>Get the hybrid identity metadata proxy resource.
      *
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
-     * @param provisionedClustersName Parameter for the name of the provisioned cluster.
+     * @param resourceName Parameter for the name of the provisioned cluster.
      * @param hybridIdentityMetadataResourceName Parameter for the name of the hybrid identity metadata resource.
      * @param context The context to associate with this operation.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
@@ -25,10 +25,7 @@ public interface HybridIdentityMetadatas {
      * @return the hybrid identity metadata proxy resource along with {@link Response}.
      */
     Response<HybridIdentityMetadata> getWithResponse(
-        String resourceGroupName,
-        String provisionedClustersName,
-        String hybridIdentityMetadataResourceName,
-        Context context);
+        String resourceGroupName, String resourceName, String hybridIdentityMetadataResourceName, Context context);
 
     /**
      * Get the hybrid identity metadata resource
@@ -36,7 +33,7 @@ public interface HybridIdentityMetadatas {
      * <p>Get the hybrid identity metadata proxy resource.
      *
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
-     * @param provisionedClustersName Parameter for the name of the provisioned cluster.
+     * @param resourceName Parameter for the name of the provisioned cluster.
      * @param hybridIdentityMetadataResourceName Parameter for the name of the hybrid identity metadata resource.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
      * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
@@ -44,7 +41,7 @@ public interface HybridIdentityMetadatas {
      * @return the hybrid identity metadata proxy resource.
      */
     HybridIdentityMetadata get(
-        String resourceGroupName, String provisionedClustersName, String hybridIdentityMetadataResourceName);
+        String resourceGroupName, String resourceName, String hybridIdentityMetadataResourceName);
 
     /**
      * Deletes the hybrid identity metadata resource
@@ -52,7 +49,7 @@ public interface HybridIdentityMetadatas {
      * <p>Deletes the hybrid identity metadata proxy resource.
      *
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
-     * @param provisionedClustersName Parameter for the name of the provisioned cluster.
+     * @param resourceName Parameter for the name of the provisioned cluster.
      * @param hybridIdentityMetadataResourceName Parameter for the name of the hybrid identity metadata resource.
      * @param context The context to associate with this operation.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
@@ -61,10 +58,7 @@ public interface HybridIdentityMetadatas {
      * @return the {@link Response}.
      */
     Response<Void> deleteWithResponse(
-        String resourceGroupName,
-        String provisionedClustersName,
-        String hybridIdentityMetadataResourceName,
-        Context context);
+        String resourceGroupName, String resourceName, String hybridIdentityMetadataResourceName, Context context);
 
     /**
      * Deletes the hybrid identity metadata resource
@@ -72,13 +66,13 @@ public interface HybridIdentityMetadatas {
      * <p>Deletes the hybrid identity metadata proxy resource.
      *
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
-     * @param provisionedClustersName Parameter for the name of the provisioned cluster.
+     * @param resourceName Parameter for the name of the provisioned cluster.
      * @param hybridIdentityMetadataResourceName Parameter for the name of the hybrid identity metadata resource.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
      * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
      */
-    void delete(String resourceGroupName, String provisionedClustersName, String hybridIdentityMetadataResourceName);
+    void delete(String resourceGroupName, String resourceName, String hybridIdentityMetadataResourceName);
 
     /**
      * Lists the hybrid identity metadata resources in a cluster
@@ -86,13 +80,13 @@ public interface HybridIdentityMetadatas {
      * <p>Lists the hybrid identity metadata proxy resource in a cluster.
      *
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
-     * @param provisionedClustersName Parameter for the name of the provisioned cluster.
+     * @param resourceName Parameter for the name of the provisioned cluster.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
      * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
      * @return list of hybridIdentityMetadata as paginated response with {@link PagedIterable}.
      */
-    PagedIterable<HybridIdentityMetadata> listByCluster(String resourceGroupName, String provisionedClustersName);
+    PagedIterable<HybridIdentityMetadata> listByCluster(String resourceGroupName, String resourceName);
 
     /**
      * Lists the hybrid identity metadata resources in a cluster
@@ -100,15 +94,14 @@ public interface HybridIdentityMetadatas {
      * <p>Lists the hybrid identity metadata proxy resource in a cluster.
      *
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
-     * @param provisionedClustersName Parameter for the name of the provisioned cluster.
+     * @param resourceName Parameter for the name of the provisioned cluster.
      * @param context The context to associate with this operation.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
      * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
      * @return list of hybridIdentityMetadata as paginated response with {@link PagedIterable}.
      */
-    PagedIterable<HybridIdentityMetadata> listByCluster(
-        String resourceGroupName, String provisionedClustersName, Context context);
+    PagedIterable<HybridIdentityMetadata> listByCluster(String resourceGroupName, String resourceName, Context context);
 
     /**
      * Get the hybrid identity metadata resource