+++ sdk/tables/azure-data-tables/src/test/java/com/azure/data/tables/TableClientTest.java
@@ -8,6 +8,7 @@ import com.azure.core.http.policy.ExponentialBackoff;
 import com.azure.core.http.policy.HttpLogDetailLevel;
 import com.azure.core.http.policy.HttpLogOptions;
 import com.azure.core.http.policy.RetryPolicy;
+import com.azure.core.http.rest.PagedIterable;
 import com.azure.core.http.rest.PagedResponse;
 import com.azure.core.http.rest.Response;
 import com.azure.core.test.http.AssertingHttpClientBuilder;
@@ -31,6 +32,8 @@ import com.azure.data.tables.sas.TableSasProtocol;
 import com.azure.data.tables.sas.TableSasSignatureValues;
 import com.azure.identity.ClientSecretCredential;
 import com.azure.identity.ClientSecretCredentialBuilder;
+
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Assumptions;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
@@ -1143,6 +1146,28 @@ public class TableClientTest extends TableClientTestBase {
         }
     }
 
-    // Testing for correct errors thrown
+    @Test
+    public void allowsCreationOfEntityWithEmptyStringPrimaryKey() {
+        Assertions.assertDoesNotThrow(() -> {
+            TableEntity entity = new TableEntity("", "");
+            tableClient.createEntity(entity);
+        });
+    }
 
+    @Test
+    public void allowListEntitiesWithEmptyPrimaryKey() {
+        TableEntity entity = new TableEntity("", "");
+        String entityName = testResourceNamer.randomName("name", 10);
+        entity.addProperty("Name", entityName);
+        tableClient.createEntity(entity);
+        ListEntitiesOptions options = new ListEntitiesOptions();
+        options.setFilter("PartitionKey eq '' and RowKey eq ''");
+        PagedIterable<TableEntity> response = tableClient.listEntities(options, Duration.ofSeconds(10), null);
+        ArrayList<TableEntity> responseArray = new ArrayList<TableEntity>();
+        for (TableEntity responseEntity : response) {
+            responseArray.add(responseEntity);
+        }
+        assertEquals(1, responseArray.size());
+        assertEquals(entityName, responseArray.get(0).getProperty("Name"));
+    }
 }