+++ sdk/identity/azure-identity/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.9.0-beta.2 (Unreleased)
 
 ### Features Added
+- Added CAE support to service principal authentication.
 
 ### Breaking Changes